Switching RSU query loop to run asynchronously
